Move folder getListUserForAdmin in model
